+++ pywren_ibm_cloud/compute/backends/localhost/localhost.py
@@ -25,13 +25,14 @@ class LocalhostBackend:
         self.name = 'local'
         self.queue = multiprocessing.Queue()
         self.run_dir = LOCAL_RUN_DIR
+        self.workers = self.config['workers']
 
-        for cpu in range(self.config['cores']):
+        for cpu in range(self.workers):
             p = multiprocessing.Process(target=self._process_runner)
             p.daemon = True
             p.start()
 
-        log_msg = 'PyWren v{} init for Localhost - Total workers: {} (CPUs)'.format(__version__, self.config['cores'])
+        log_msg = 'PyWren v{} init for Localhost - Total workers: {}'.format(__version__, self.workers)
         logger.info(log_msg)
         if not self.log_level:
             print(log_msg)
@@ -123,6 +124,7 @@ class LocalhostBackend:
         Runtime keys are used to uniquely identify runtimes within the storage,
         in order to know what runtimes are installed and what not.
         """
-        runtime_key = '{}_{}'.format(runtime_name, str(runtime_memory))
+        runtime_key = '{}_{}MB'.format(runtime_name, str(runtime_memory))
+        runtime_key = os.path.join(self.name, runtime_key)
 
         return runtime_key